Keep track of loading state in RunsListViewController.isLoadingData
That var was there but unused. This isn't quite right (it may do the wrong thing if there are concurrent requests)
But probably OK for now. In the future, a refactored API client might want to return something that tracks the request state and allows cancellation. And RunsListController might want to enforce only making a single reaches request at a time.