Remove an alert presented if AddRiverFlowTableViewController had no selected run
Change selectedRun to an implicitly unwrapped optional instead.
This controller makes no sense without a selectedRun and we should just crash in that case, and fix whatever case allowed that to happen.

Also add shouldPerformSegue(withIdentifier:) on RiverFlowsViewController to guard against seguing if there's no run. But it shouldn't be optional here either...