Another try: show loading cell or 'no rivers available' cell as appropriate
Rule: if there are any reaches stored locally, show the empty cell if the current filters & search don't have any results.
Show the loading cell if there are no reaches stored locally.

This is still not great -- it's a proxy for what we really want to know:
if there are no results, is there also a relevant network request in progress, that might change that?

But answering that is harder and would need more API refactor. This is hopefully mostly right most of the time.

Fixes #206