Change AWGQLApiHelper's error callback type from (Error?, String?) to (Error)
The old type had either an Error or a string containing an error message.
That second param is now an Error type (Errors.graphQLError) with a message attached

In the future it would be great to go through and create more precise errors for some of these.
For now, not super important as they pretty much all end up just getting print()ed or possibly used in a toast message.