Refactor API.getAlertsForReach() and the stuff that uses it
- Returns array of new Alert model (not the generated type)
- RunAlertsViewController is the consumer, refactor to use Alert

Also includes conversion to/from a [String:String] dict, to maintain compatibility and continue the alert storage in DefaultsManager.reachAlerts.
I almost ripped this out, but I'm not sure why it's there, so leaving it for now. But we should pull it and replace it if needed. UserDefaults is intended for small key value data, not Dictionary of unbounded size.