Refactor filters (#243)
* Add a Filters struct with same properties as the filtering bits of DefaultsManager

And add a getter/setter to DefaultsManager for it.

* Extract predicate generation from RunsListViewController to Filters

Provide Filters.combinedPredicate() to get a compound predicate for a set of filters and use that in runs list

* Remove predicate stuff from MapViewController, use Filters instead

Note that this changes the behavior of the map:

- The class filter was ignored and is now applied. (This seems like a bug previously)
- The runnable filter was ignored and is now applied, but it's only settable from the runs list controller. See FIXME comment in commit

* Make a static default filter to set on first run

* Use DefaultsManager.shared.filters to manipulate filters in OnboardLocationViewController

* Maintain FilterViewController state while editing in a Filters var, and save it on dismiss

Done with minimal refactor beyond what's needed to adopt Filters

* Mark backing store of filter properties in DefaultsManager private

All access of filter properties now uses Filters

* More refactor of FilterViewController

User visible changes:
- Reload regions table when the clear button is pressed, so it's not out of sync
- Display selected regions alphabetically
- Update the distance filter label as the slider is changed so you don't have to let go to see your filtered distance

* Reorganize FilterViewController file

Split protocol implementations into extensions. Just reordering, no other changes.

* Use xmark.circle (SF Symbol) for clear region filters button in FilterViewController

Note that this will not work right on <iOS 13 ... there are a number of other SF Symbols already in use with the same issue.

* Not necessary to reload the filter controller's collection view every time the distance slider changes

I left this in when switching from touchUpInside to valueChanged on the slider --
but it creates some visual jank, and makes the geocoder thrash and eventually start failing.

The only state that's changing is filters.distanceFilter and the distance filter label, and those are both updated.

* Post a notification when DefaultsManager.filters change

To give other parts of the app a way to find out that happened

* Change map view fetched results controller lifecycle:

- Create and do initial fetch on viewDidLoad, not viewDidAppear
- Don't discard fetchedResultsController on viewWillDisappear, keep it around
- Watch for filter changes, re-fetch and re-zoom when that happens

Fixes #231.

* Another try: show loading cell or 'no rivers available' cell as appropriate

Rule: if there are any reaches stored locally, show the empty cell if the current filters & search don't have any results.
Show the loading cell if there are no reaches stored locally.

This is still not great -- it's a proxy for what we really want to know:
if there are no results, is there also a relevant network request in progress, that might change that?

But answering that is harder and would need more API refactor. This is hopefully mostly right most of the time.

Fixes #206

* Remove generated header

* Remove gratuitous extensions in Filters