Remove error alerts from GalleryViewController
These were there to expose things like not having a photo url to the user as an error when they tried to expand the photo
Instead, selecting a photo should construct its URL, and fail if it does. And the expand button should be hidden if it can't be used successfully

I ended up doing quite a bit of refactor to support those changes.

This controller still has lots of issues and needs more work:
- Its data is provided as [String: String?], awkward to work with
- There's no photo caching so it won't work offline
- The initial photo is set from a different source as the rest, and that leaves the caption/author/etc unavailable so they show confusing default strings