Create a type GaugeDataPoint and use it to represent gauge data
One issue with this commit is, there's already a GaugeObservation type, used in RiverFlowsViewController.
They do more or less the same thing, but GaugeObservation has some other properties, is currently retrieved via GraphQL and has more optionals. It also has a [Photo] property, since it's used for posted observations.
So, I'm not sure if these should be the same or not. Maybe a Gauge Observation should have a GaugeDataPoint with it's actual observation, and layer on the other stuff it represents. In any case, seems like they should be unified one way or another.

Also, this uses Codable/JSONDecoder to deserialize, rather than SwiftyJSON. Note the custom Decodable implementation, needed to work around some response behavior:

- reading (the gauge value) is returned as a string
- updated (the timestamp) is sometimes returned as a number, sometimes as a string