Define standard and scroll edge appearance for the app's main tab bar
Opaque, white background for now. In iOS 15, the standard is to use a 'system material' which blurs the underlying content -- that would look less dated. However, we'd need to revise or change the tab bar icons to support this, they look pretty weird right now if they aren't on a white background.

Note that this could also be done via the appearance API, like the app's navigation bars. But, since there's only one tab bar in the app it's more straightforward to set it up in the storyboard.